+++ NEWS.md
@@ -5,7 +5,8 @@
 * Improve message feedback for downloading objects
 * `gcs_upload()` will now retry with a warning if using legacy ACL settings for a bucket with uniform bucket-level access enabled - will retry with `predefinedAcl='bucketLevel'`
 * Add `gcs_parse_rds()` to help parsing RDS objects for `gcs_get_object()` #146
-* Add ability to request noncurrent versions of objects in `gcs_get_object()` via `generation` parameter.
+* Add ability to request non-current versions of objects in `gcs_get_object()` via `generation` parameter.
+* Add `gcs_delete_bucket_objects()` to help delete all objects from a bucket, including older object generation versions (#152)
 
 # googleCloudStorageR 0.6.0